+++ client/src/api/_liberationApi.ts
@@ -425,6 +425,7 @@ export type Tgo = {
   threat_ranges: number[];
   detection_ranges: number[];
   dead: boolean;
+  purchasable: boolean;
   sidc: string;
   task?: string[];
 };